\name{glmfunk-package}
\alias{glmfunk-package}
\alias{glmfunk}
\docType{package}
\title{
\packageTitle{glmfunk}
}
\description{
\packageDescription{glmfunk}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{glmfunk}
\packageIndices{glmfunk}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{glmfunk}

Maintainer: \packageMaintainer{glmfunk}
}
\references{
~~ Literature or other references for background information ~~
}

\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
